+++ src/tquality_core/utils/pydantic_utils/__init__.py
@@ -1,0 +1,4 @@
+from tquality_core.utils.pydantic_utils._model_metadata import ModelMetadata
+from tquality_core.utils.pydantic_utils._settings_metadata import SettingsMetadata
+
+__all__ = ["ModelMetadata", "SettingsMetadata"]